+++ src/main/java/io/apercova/quickcli/Command.java
@@ -1,5 +1,8 @@
 package io.apercova.quickcli;
 
+import io.apercova.quickcli.exception.ExecutionException;
+import io.apercova.quickcli.annotation.CLIArgument;
+import io.apercova.quickcli.annotation.CLICommand;
 import java.io.Closeable;
 import java.io.IOException;
 import java.io.PrintStream;
