+++ src/main/java/io/apercova/quickcli/annotation/CLIDatatypeConverter.java
@@ -1,5 +1,6 @@
-package io.apercova.quickcli;
+package io.apercova.quickcli.annotation;
 
+import io.apercova.quickcli.DatatypeConverter;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
