+++ src/test/java/io/apercova/quickcli/BindExamplesTest.java
